+++ profile-repository-impl/src/main/java/com/dwarfeng/prorepo/impl/bean/HibernateMapper.java
@@ -1,0 +1,37 @@
+package com.dwarfeng.prorepo.impl.bean;
+
+import com.dwarfeng.prorepo.impl.bean.entity.HibernateProfile;
+import com.dwarfeng.prorepo.stack.bean.entity.Profile;
+import com.dwarfeng.subgrade.sdk.bean.key.HibernateLongIdKey;
+import com.dwarfeng.subgrade.sdk.bean.key.HibernateStringIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import org.mapstruct.InheritInverseConfiguration;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+
+/**
+ * Hibernate Bean 映射器。
+ *
+ * @author DwArFeng
+ * @since 1.1.0
+ */
+@Mapper
+public interface HibernateMapper {
+
+    HibernateLongIdKey longIdKeyToHibernate(LongIdKey longIdKey);
+
+    @InheritInverseConfiguration
+    LongIdKey longIdKeyFromHibernate(HibernateLongIdKey hibernateLongIdKey);
+
+    HibernateStringIdKey stringIdKeyToHibernate(StringIdKey stringIdKey);
+
+    @InheritInverseConfiguration
+    StringIdKey stringIdKeyFromHibernate(HibernateStringIdKey hibernateStringIdKey);
+
+    @Mapping(target = "stringId", ignore = true)
+    HibernateProfile profileToHibernate(Profile profile);
+
+    @InheritInverseConfiguration
+    Profile profileFromHibernate(HibernateProfile hibernateProfile);
+}