+++ profile-repository-sdk/src/main/java/com/dwarfeng/prorepo/sdk/bean/FastJsonMapper.java
@@ -1,0 +1,35 @@
+package com.dwarfeng.prorepo.sdk.bean;
+
+import com.dwarfeng.prorepo.sdk.bean.entity.FastJsonProfile;
+import com.dwarfeng.prorepo.stack.bean.entity.Profile;
+import com.dwarfeng.subgrade.sdk.bean.key.FastJsonLongIdKey;
+import com.dwarfeng.subgrade.sdk.bean.key.FastJsonStringIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import org.mapstruct.InheritInverseConfiguration;
+import org.mapstruct.Mapper;
+
+/**
+ * FastJson Bean 映射器。
+ *
+ * @author DwArFeng
+ * @since 1.1.0
+ */
+@Mapper
+public interface FastJsonMapper {
+
+    FastJsonLongIdKey longIdKeyToFastJson(LongIdKey longIdKey);
+
+    @InheritInverseConfiguration
+    LongIdKey longIdKeyFromFastJson(FastJsonLongIdKey fastJsonLongIdKey);
+
+    FastJsonStringIdKey stringIdKeyToFastJson(StringIdKey stringIdKey);
+
+    @InheritInverseConfiguration
+    StringIdKey stringIdKeyFromFastJson(FastJsonStringIdKey fastJsonStringIdKey);
+
+    FastJsonProfile profileToFastJson(Profile profile);
+
+    @InheritInverseConfiguration
+    Profile profileFromFastJson(FastJsonProfile fastJsonProfile);
+}